Scoring script for German narratives
The following script allows for quick and flexible translation and subsequent scoring of non-English texts with the Autobiographical Interview (Levene, 2005) heavily relying on the automated-scoring approach developed and provided by van Genugten & Schacter (2024).  
It is a generalization and extension of the code used by Annika Kuelpmann for this project: https://osf.io/vcns4.  
The liscence for this script is:   
The authors of this script are: Giuliano Groer (https://orcid.org/0009-0009-2656-3796), Annika Kuelpmann (https://orcid.org/0000-0002-0256-2037), Laurin Plank (https://orcid.org/0000-0002-8846-5405)